[DA,JS,CH] refactored tests for filtering of owned ingredients and the recipe class
